Move benchmarks.py to model_systems
